requirements.txt
fastapi==0.63.0
python-telegram-bot==20.6
python-dotenv==1.0.0
mongoengine==0.27.0